make CSVFeatureLogger class of its own
